handle EOF / C-d gracefully at prompt
The current behavior is to enter an infinite loop when C-d is pressed
(or in any other way an EOF is received by std::cin). This is because
std::cin gets into a "closed" state, so the next time it is queried it
does not wait for user input but instead immediately returns an empty
string.

This commit simply clears this state from std::cin to avoid the infinite
loop behavior. The effect of pressing C-d without any input at the
prompt is now the same as that of just pressing Enter.